refactor(credential): purge dead code, delete rotation, fix error classification, decompose facade + resolver
Major cleanup pass on nebula-credential toward Tokio-grade quality, driven by a
13-area audit (42 confirmed findings). ~6k LOC net removed; five god-file
concerns addressed. Verified green throughout: 420 tests, clippy -D warnings
(default + rotation feature), rustdoc -D warnings.

Dead code + duplication (verified zero-consumer):
- Delete legacy #[derive(Credential)] macro (730 LOC, had a todo!() panic path),
  StaticProtocol, ext.rs/HasCredentialsExt, runtime/scoped_accessor.rs (engine
  ships a duplicate accessor), and an obsolete trybuild probe.
- Delete both dead rotation trees (src/rotation/ + src/runtime/rotation/,
  ~4,665 LOC, feature-gated with no dependent enabling it) + storage
  RotationBackup + the engine smoke test + record's RotationPolicy field. The
  OAuth2 try_oauth2_refresh C::KEY relic stays feature-gated (DESIGN Phase-3
  boundary); resource's rotation feature (the live credential fan-out) is
  untouched.
- Collapse SchemeKind/SchemeIdentity (a rename-duplicate of
  nebula_core::auth::AuthPattern) into a name-carrying SchemeMismatch.

Error-classification correctness (3 confirmed P1 bugs):
- resolve_error_to_credential_error is now permanence-aware: permanent failures
  (reauth, corrupt state, not-found) map to non-retryable variants instead of a
  blanket retryable ServerError, so a retry-driven caller no longer hammers the
  IdP or loops forever.
- New CredentialServiceError::ReauthRequired{credential_id, reason}
  (category=validation -> 401), mapped at the facade, ops, and api layers. Was
  flattened to Internal/500; the validation category also stops the reauth
  retry-hammer.
- executor_error_to_service_error replaces the blanket
  ExecutorError -> ValidationFailed/400 (timeout -> 503, store outage -> 500,
  expired/consumed pending token -> 401).

Architecture (behaviour-preserving decomposition):
- facade.rs 1542 -> 408 LOC (-73%): slot / acquire / discovery / crud /
  capabilities extracted into cohesive service/*.rs `impl CredentialService`
  modules.
- resolver.rs 1922 -> 1591 LOC: resolution error taxonomy + fail-closed
  owner/tombstone gates extracted to runtime/resolve_error.rs.
