+++ crates/credential/Cargo.toml
@@ -10,7 +10,7 @@ repository.workspace = true
 
 [dependencies]
 # Tokio features:
-#   `time`   — `timeout` in rotation/error; device-code polling paths
+#   `time`   — `timeout` on refresh/resolve paths; device-code polling paths
 #   `macros` — #[tokio::test]
 #   `rt`     — #[tokio::test] expansion; `tokio::spawn` for the lease lifecycle task
 #   `sync`   — `mpsc`/`oneshot` channels driving the lease-lifecycle scheduler
@@ -124,6 +124,11 @@ trybuild = "1.0"
 
 [features]
 default = []
+# Transitional: gates the OAuth2 HTTP-refresh path in `runtime::resolver`
+# (`try_oauth2_refresh`, the `C::KEY == OAuth2Credential::KEY` special-case).
+# The rotation subsystem it was named for was deleted; this relic is removed
+# when OAuth2 refresh moves onto its own grant-discriminant `Refreshable::refresh`
+# (DESIGN §17 Phase 3). Kept feature-gated so it stays out of default builds.
 rotation = []
 
 [package.metadata.docs.rs]
