--- crates/credential/src/rotation/events.rs
@@ -1,828 +1,0 @@
-//! Rotation Notification Events
-//!
-//! Provides event types and notification abstraction for rotation lifecycle events.
-
-use std::future::Future;
-
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-
-use super::error::RotationResult;
-use crate::CredentialId;
-
-/// Rollback data (boxed to reduce enum size)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RollbackData {
-    /// Credential that had rotation rolled back
-    pub credential_id: CredentialId,
-
-    /// When rollback occurred
-    pub rolled_back_at: DateTime<Utc>,
-
-    /// Rotation transaction ID
-    pub transaction_id: String,
-
-    /// Reason for rollback
-    pub reason: String,
-
-    /// Error classification (transient, permanent, unknown)
-    pub error_classification: Option<String>,
-
-    /// Number of retries attempted before rollback
-    pub retry_count: u32,
-
-    /// Rotation state when rollback was triggered
-    pub rotation_state: String,
-}
-
-/// Emergency rotation data (boxed to reduce enum size)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EmergencyRotationData {
-    /// Credential being rotated
-    pub credential_id: CredentialId,
-
-    /// When emergency was triggered
-    pub triggered_at: DateTime<Utc>,
-
-    /// Rotation transaction ID
-    pub transaction_id: String,
-
-    /// Reason for emergency rotation
-    pub reason: String,
-
-    /// Who triggered the emergency rotation
-    pub triggered_by: String,
-
-    /// Incident tracking ID (if linked)
-    pub incident_id: Option<String>,
-
-    /// Whether old credential was immediately revoked
-    pub immediate_revoke: bool,
-}
-
-/// Emitted after every successful credential rotation.
-///
-/// `ResourceManager` subscribes to this to trigger pool re-authorization.
-/// Carries only `credential_id` + `generation` — never credential state,
-/// which would leak secrets to any EventBus subscriber that logs events.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CredentialRotationEvent {
-    /// Which credential instance was rotated.
-    pub credential_id: CredentialId,
-    /// Monotonically increasing generation counter for this credential.
-    /// Subscribers use this to detect stale re-authorization attempts.
-    pub generation: u64,
-}
-
-/// Notification event for rotation lifecycle
-///
-/// Events are emitted at key points during rotation:
-/// - Scheduled: When rotation is scheduled (with advance notice)
-/// - Starting: Just before rotation begins
-/// - Complete: After successful rotation
-/// - Failed: If rotation fails
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-#[non_exhaustive]
-pub enum NotificationEvent {
-    /// Rotation has been scheduled
-    RotationScheduled {
-        /// Credential being rotated
-        credential_id: CredentialId,
-
-        /// When rotation will occur
-        scheduled_at: DateTime<Utc>,
-
-        /// Time remaining until rotation
-        time_until: std::time::Duration,
-    },
-
-    /// Rotation is about to start
-    RotationStarting {
-        /// Credential being rotated
-        credential_id: CredentialId,
-
-        /// Current timestamp
-        starting_at: DateTime<Utc>,
-
-        /// Rotation transaction ID
-        transaction_id: String,
-    },
-
-    /// Rotation completed successfully
-    RotationComplete {
-        /// Credential that was rotated
-        credential_id: CredentialId,
-
-        /// When rotation completed
-        completed_at: DateTime<Utc>,
-
-        /// Rotation transaction ID
-        transaction_id: String,
-
-        /// Duration of rotation process
-        duration: std::time::Duration,
-
-        /// Old version number
-        old_version: u32,
-
-        /// New version number
-        new_version: u32,
-    },
-
-    /// Rotation failed
-    RotationFailed {
-        /// Credential that failed rotation
-        credential_id: CredentialId,
-
-        /// When rotation failed
-        failed_at: DateTime<Utc>,
-
-        /// Rotation transaction ID
-        transaction_id: String,
-
-        /// Error message
-        error: String,
-
-        /// Retry attempt number (0 for first attempt)
-        retry_attempt: u32,
-    },
-
-    /// Emergency manual rotation triggered (security incident)
-    EmergencyRotation(Box<EmergencyRotationData>),
-
-    /// Rotation was rolled back due to failure
-    ///
-    /// # T077: Rollback Event Logging
-    RotationRolledBack(Box<RollbackData>),
-}
-
-impl NotificationEvent {
-    /// Get the credential ID associated with this event
-    pub fn credential_id(&self) -> &CredentialId {
-        match self {
-            NotificationEvent::RotationScheduled { credential_id, .. } => credential_id,
-            NotificationEvent::RotationStarting { credential_id, .. } => credential_id,
-            NotificationEvent::RotationComplete { credential_id, .. } => credential_id,
-            NotificationEvent::RotationFailed { credential_id, .. } => credential_id,
-            NotificationEvent::EmergencyRotation(data) => &data.credential_id,
-            NotificationEvent::RotationRolledBack(data) => &data.credential_id,
-        }
-    }
-
-    /// Get the event timestamp
-    pub fn timestamp(&self) -> DateTime<Utc> {
-        match self {
-            NotificationEvent::RotationScheduled { scheduled_at, .. } => *scheduled_at,
-            NotificationEvent::RotationStarting { starting_at, .. } => *starting_at,
-            NotificationEvent::RotationComplete { completed_at, .. } => *completed_at,
-            NotificationEvent::RotationFailed { failed_at, .. } => *failed_at,
-            NotificationEvent::EmergencyRotation(data) => data.triggered_at,
-            NotificationEvent::RotationRolledBack(data) => data.rolled_back_at,
-        }
-    }
-
-    /// Get a human-readable event description
-    pub fn description(&self) -> String {
-        match self {
-            NotificationEvent::RotationScheduled {
-                credential_id,
-                scheduled_at,
-                ..
-            } => {
-                format!("Rotation scheduled for credential {credential_id} at {scheduled_at}")
-            },
-            NotificationEvent::RotationStarting {
-                credential_id,
-                transaction_id,
-                ..
-            } => {
-                format!(
-                    "Rotation starting for credential {credential_id} (transaction: {transaction_id})"
-                )
-            },
-            NotificationEvent::RotationComplete {
-                credential_id,
-                old_version,
-                new_version,
-                ..
-            } => {
-                format!(
-                    "Rotation complete for credential {credential_id} (v{old_version} → v{new_version})"
-                )
-            },
-            NotificationEvent::RotationFailed {
-                credential_id,
-                error,
-                retry_attempt,
-                ..
-            } => {
-                format!(
-                    "Rotation failed for credential {credential_id} (attempt {}): {error}",
-                    retry_attempt + 1,
-                )
-            },
-            NotificationEvent::EmergencyRotation(data) => {
-                let revoke_status = if data.immediate_revoke {
-                    "IMMEDIATE REVOCATION"
-                } else {
-                    "with grace period"
-                };
-                let incident_info = data
-                    .incident_id
-                    .as_ref()
-                    .map(|id| format!(" [Incident: {id}]"))
-                    .unwrap_or_default();
-                format!(
-                    "🚨 EMERGENCY: {} rotation for credential {} by {} - Reason: {}{}",
-                    revoke_status,
-                    data.credential_id,
-                    data.triggered_by,
-                    data.reason,
-                    incident_info
-                )
-            },
-            NotificationEvent::RotationRolledBack(data) => {
-                let classification = data
-                    .error_classification
-                    .as_ref()
-                    .map(|c| format!(" [{c}]"))
-                    .unwrap_or_default();
-                format!(
-                    "🔄 ROLLBACK: Rotation rolled back for credential {} after {} retries in state '{}'{} - Reason: {}",
-                    data.credential_id,
-                    data.retry_count,
-                    data.rotation_state,
-                    classification,
-                    data.reason
-                )
-            },
-        }
-    }
-}
-
-/// Trait for notification senders
-///
-/// Implement this trait to send notifications via different channels:
-/// - Email (SMTP)
-/// - Slack webhooks
-/// - PagerDuty
-/// - Custom logging
-/// - Metrics/monitoring systems
-///
-/// # Example
-///
-/// ```rust
-/// use chrono::Utc;
-///
-/// use nebula_credential::CredentialId;
-/// use nebula_credential::rotation::RotationResult;
-/// use nebula_credential::rotation::events::{NotificationEvent, NotificationSender};
-///
-/// struct SlackNotifier {
-///     webhook_url: String,
-/// }
-///
-/// impl NotificationSender for SlackNotifier {
-///     async fn send(&self, event: &NotificationEvent) -> RotationResult<()> {
-///         // A production sender would POST `event.description()` to
-///         // `self.webhook_url`; this example just renders the line.
-///         let _payload = event.description();
-///         let _ = &self.webhook_url;
-///         Ok(())
-///     }
-/// }
-///
-/// let notifier = SlackNotifier {
-///     webhook_url: "https://hooks.example/abc".to_string(),
-/// };
-/// let event = NotificationEvent::RotationStarting {
-///     credential_id: CredentialId::new(),
-///     starting_at: Utc::now(),
-///     transaction_id: "tx-1".to_string(),
-/// };
-///
-/// let runtime = tokio::runtime::Builder::new_current_thread()
-///     .enable_all()
-///     .build()
-///     .unwrap();
-/// runtime.block_on(async {
-///     notifier.send(&event).await.unwrap();
-/// });
-/// ```
-pub trait NotificationSender: Send + Sync {
-    /// Send a notification event
-    ///
-    /// # Arguments
-    ///
-    /// * `event` - The notification event to send
-    ///
-    /// # Returns
-    ///
-    /// * `RotationResult<()>` - Success or error
-    fn send(&self, event: &NotificationEvent) -> impl Future<Output = RotationResult<()>> + Send;
-}
-
-/// Send notification with retry logic
-///
-/// Attempts to send notification with exponential backoff retry on failure.
-///
-/// # Arguments
-///
-/// * `sender` - Notification sender implementation
-/// * `event` - Event to send
-/// * `policy` - Retry policy configuration
-///
-/// # Returns
-///
-/// * `RotationResult<()>` - Success or final error after retries
-///
-/// # Example
-///
-/// ```rust
-/// use chrono::Utc;
-///
-/// use nebula_credential::CredentialId;
-/// use nebula_credential::rotation::RotationResult;
-/// use nebula_credential::rotation::events::{
-///     NotificationEvent, NotificationSender, send_notification,
-/// };
-/// use nebula_resilience::retry::{BackoffConfig, JitterConfig, RetryConfig};
-///
-/// struct NullNotifier;
-///
-/// impl NotificationSender for NullNotifier {
-///     async fn send(&self, _event: &NotificationEvent) -> RotationResult<()> {
-///         Ok(())
-///     }
-/// }
-///
-/// let config = RetryConfig::new(6)
-///     .unwrap()
-///     .backoff(BackoffConfig::exponential_default())
-///     .jitter(JitterConfig::Full { factor: 0.25, seed: None });
-///
-/// let event = NotificationEvent::RotationComplete {
-///     credential_id: CredentialId::new(),
-///     completed_at: Utc::now(),
-///     transaction_id: "tx-42".to_string(),
-///     duration: std::time::Duration::from_secs(30),
-///     old_version: 1,
-///     new_version: 2,
-/// };
-///
-/// let runtime = tokio::runtime::Builder::new_current_thread()
-///     .enable_all()
-///     .build()
-///     .unwrap();
-/// runtime.block_on(async {
-///     send_notification(&NullNotifier, &event, config).await.unwrap();
-/// });
-/// ```
-pub async fn send_notification<S: NotificationSender>(
-    sender: &S,
-    event: &NotificationEvent,
-    config: nebula_resilience::retry::RetryConfig<super::error::RotationError>,
-) -> RotationResult<()> {
-    let config = config.on_retry(
-        |err: &super::error::RotationError, delay: std::time::Duration, attempt: u32| {
-            tracing::warn!(
-                attempt,
-                delay_ms = delay.as_millis(),
-                error = %err,
-                "Notification send failed, retrying after delay"
-            );
-        },
-    );
-
-    nebula_resilience::retry_with(config, || async { sender.send(event).await })
-        .await
-        .map_err(|call_err| match call_err {
-            nebula_resilience::CallError::Operation(e)
-            | nebula_resilience::CallError::RetriesExhausted { last: e, .. } => e,
-            _ => unreachable!("retry_with only returns Operation or RetriesExhausted"),
-        })
-}
-
-/// Log a rollback event with structured logging
-///
-/// # T077: Log Rollback Event
-///
-/// Creates a structured log entry and notification event for rollback operations.
-///
-/// # Arguments
-///
-/// * `error_log` - Detailed error log from the rotation failure
-/// * `sender` - Optional notification sender for alerting
-/// * `policy` - Retry policy for notification delivery
-///
-/// # Example
-///
-/// ```rust
-/// use nebula_credential::CredentialId;
-/// use nebula_credential::rotation::RotationResult;
-/// use nebula_credential::rotation::error::RotationErrorLog;
-/// use nebula_credential::rotation::events::{
-///     NotificationEvent, NotificationSender, log_rollback_event,
-/// };
-/// use nebula_resilience::retry::{BackoffConfig, RetryConfig};
-///
-/// struct NullNotifier;
-///
-/// impl NotificationSender for NullNotifier {
-///     async fn send(&self, _event: &NotificationEvent) -> RotationResult<()> {
-///         Ok(())
-///     }
-/// }
-///
-/// let error_log = RotationErrorLog::new(
-///     "tx-9c2",
-///     CredentialId::new(),
-///     "Validation failed",
-/// )
-/// .with_rollback_triggered()
-/// .with_retry_count(3);
-///
-/// let config = RetryConfig::new(3)
-///     .unwrap()
-///     .backoff(BackoffConfig::exponential_default());
-///
-/// let runtime = tokio::runtime::Builder::new_current_thread()
-///     .enable_all()
-///     .build()
-///     .unwrap();
-/// runtime.block_on(async {
-///     log_rollback_event(&error_log, Some(&NullNotifier), config)
-///         .await
-///         .unwrap();
-/// });
-/// ```
-pub async fn log_rollback_event<S: NotificationSender>(
-    error_log: &super::error::RotationErrorLog,
-    sender: Option<&S>,
-    config: nebula_resilience::retry::RetryConfig<super::error::RotationError>,
-) -> RotationResult<()> {
-    use tracing::warn;
-
-    // Log to tracing system
-    warn!(
-        transaction_id = %error_log.transaction_id,
-        credential_id = %error_log.credential_id,
-        retry_count = error_log.retry_count,
-        error_classification = ?error_log.error_classification,
-        rotation_state = ?error_log.rotation_state,
-        "Rotation rolled back due to failure: {}",
-        error_log.error_message
-    );
-
-    // Send notification if sender provided
-    if let Some(sender) = sender {
-        let event = NotificationEvent::RotationRolledBack(Box::new(RollbackData {
-            credential_id: error_log.credential_id,
-            rolled_back_at: error_log.occurred_at,
-            transaction_id: error_log.transaction_id.clone(),
-            reason: error_log.error_message.clone(),
-            error_classification: error_log.error_classification.clone(),
-            retry_count: error_log.retry_count,
-            rotation_state: error_log
-                .rotation_state
-                .clone()
-                .unwrap_or_else(|| "unknown".to_string()),
-        }));
-
-        send_notification(sender, &event, config).await?;
-    }
-
-    Ok(())
-}
-
-/// Transaction audit log entry
-///
-/// # T094: TransactionLog for Audit Trail
-///
-/// Records all state transitions and events during a rotation transaction
-/// for audit, compliance, and debugging purposes.
-///
-/// # Example
-///
-/// ```rust
-/// use nebula_credential::CredentialId;
-/// use nebula_credential::rotation::events::TransactionLog;
-/// use nebula_credential::rotation::state::RotationState;
-///
-/// let mut log = TransactionLog::new("tx-001".to_string(), CredentialId::new());
-///
-/// // Log a state transition.
-/// log.log_transition(RotationState::Creating, "Starting credential creation");
-///
-/// // Log a validation result.
-/// log.log_validation_result(true, "Connection test passed");
-///
-/// assert_eq!(log.entry_count(), 2);
-/// assert!(!log.has_errors());
-/// assert!(log.get_error_entries().is_empty());
-/// ```
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TransactionLog {
-    /// Transaction ID this log belongs to
-    pub transaction_id: String,
-
-    /// Credential ID being rotated
-    pub credential_id: CredentialId,
-
-    /// Log entries (chronological order)
-    pub entries: Vec<TransactionLogEntry>,
-
-    /// When transaction started
-    pub started_at: DateTime<Utc>,
-
-    /// When transaction completed (if finished)
-    pub completed_at: Option<DateTime<Utc>>,
-
-    /// Final transaction outcome
-    pub outcome: Option<TransactionOutcome>,
-}
-
-/// Individual log entry in transaction audit trail
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TransactionLogEntry {
-    /// When this entry was created
-    pub timestamp: DateTime<Utc>,
-
-    /// Entry type
-    pub entry_type: LogEntryType,
-
-    /// Entry message
-    pub message: String,
-
-    /// Additional context (JSON-serializable)
-    pub context: Option<serde_json::Value>,
-}
-
-/// Type of log entry
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-#[non_exhaustive]
-pub enum LogEntryType {
-    /// State transition occurred
-    StateTransition,
-
-    /// Validation was performed
-    Validation,
-
-    /// Error occurred
-    Error,
-
-    /// Warning issued
-    Warning,
-
-    /// Informational message
-    Info,
-
-    /// Rollback triggered
-    Rollback,
-
-    /// Commit succeeded
-    Commit,
-}
-
-/// Final outcome of transaction
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-#[non_exhaustive]
-pub enum TransactionOutcome {
-    /// Transaction committed successfully
-    Committed,
-
-    /// Transaction was rolled back
-    RolledBack,
-
-    /// Transaction was aborted
-    Aborted,
-
-    /// Transaction is still in progress
-    InProgress,
-}
-
-impl TransactionLog {
-    /// Create a new transaction log
-    pub fn new(transaction_id: String, credential_id: CredentialId) -> Self {
-        Self {
-            transaction_id,
-            credential_id,
-            entries: Vec::new(),
-            started_at: Utc::now(),
-            completed_at: None,
-            outcome: None,
-        }
-    }
-
-    /// Add a log entry
-    fn add_entry(&mut self, entry_type: LogEntryType, message: impl Into<String>) {
-        self.entries.push(TransactionLogEntry {
-            timestamp: Utc::now(),
-            entry_type,
-            message: message.into(),
-            context: None,
-        });
-    }
-
-    /// Add a log entry with context
-    fn add_entry_with_context(
-        &mut self,
-        entry_type: LogEntryType,
-        message: impl Into<String>,
-        context: serde_json::Value,
-    ) {
-        self.entries.push(TransactionLogEntry {
-            timestamp: Utc::now(),
-            entry_type,
-            message: message.into(),
-            context: Some(context),
-        });
-    }
-
-    /// Log a state transition
-    pub fn log_transition(
-        &mut self,
-        new_state: super::state::RotationState,
-        message: impl Into<String>,
-    ) {
-        let context = serde_json::json!({
-            "new_state": format!("{:?}", new_state),
-        });
-        self.add_entry_with_context(LogEntryType::StateTransition, message, context);
-    }
-
-    /// Log a validation result
-    pub fn log_validation_result(&mut self, passed: bool, message: impl Into<String>) {
-        let context = serde_json::json!({
-            "passed": passed,
-        });
-        self.add_entry_with_context(LogEntryType::Validation, message, context);
-    }
-
-    /// Log an error
-    pub fn log_error(&mut self, message: impl Into<String>) {
-        self.add_entry(LogEntryType::Error, message);
-    }
-
-    /// Log a warning
-    pub fn log_warning(&mut self, message: impl Into<String>) {
-        self.add_entry(LogEntryType::Warning, message);
-    }
-
-    /// Log info message
-    pub fn log_info(&mut self, message: impl Into<String>) {
-        self.add_entry(LogEntryType::Info, message);
-    }
-
-    /// Log rollback event
-    pub fn log_rollback(&mut self, reason: impl Into<String>) {
-        self.add_entry(LogEntryType::Rollback, reason);
-        self.completed_at = Some(Utc::now());
-        self.outcome = Some(TransactionOutcome::RolledBack);
-    }
-
-    /// Log commit event
-    pub fn log_commit(&mut self) {
-        self.add_entry(LogEntryType::Commit, "Transaction committed successfully");
-        self.completed_at = Some(Utc::now());
-        self.outcome = Some(TransactionOutcome::Committed);
-    }
-
-    /// Check if log contains any errors
-    pub fn has_errors(&self) -> bool {
-        self.entries
-            .iter()
-            .any(|e| e.entry_type == LogEntryType::Error)
-    }
-
-    /// Get all error entries
-    pub fn get_error_entries(&self) -> Vec<&TransactionLogEntry> {
-        self.entries
-            .iter()
-            .filter(|e| e.entry_type == LogEntryType::Error)
-            .collect()
-    }
-
-    /// Get transaction duration (if completed)
-    pub fn duration(&self) -> Option<chrono::Duration> {
-        self.completed_at.map(|end| end - self.started_at)
-    }
-
-    /// Get total entry count
-    pub fn entry_count(&self) -> usize {
-        self.entries.len()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn test_cred_id() -> CredentialId {
-        // Deterministic fixture — avoids random IDs in assertions / snapshot-style output.
-        "cred_01ARZ3NDEKTSV4RRFFQ69G5FAV"
-            .parse()
-            .expect("static test CredentialId must parse")
-    }
-
-    #[test]
-    fn test_notification_event_credential_id() {
-        let cred_id = test_cred_id();
-
-        let event = NotificationEvent::RotationScheduled {
-            credential_id: cred_id,
-            scheduled_at: Utc::now(),
-            time_until: std::time::Duration::from_hours(1),
-        };
-
-        assert_eq!(event.credential_id(), &cred_id);
-    }
-
-    #[test]
-    fn test_notification_event_description() {
-        let cred_id = test_cred_id();
-
-        let event = NotificationEvent::RotationComplete {
-            credential_id: cred_id,
-            completed_at: Utc::now(),
-            transaction_id: "tx-123".to_string(),
-            duration: std::time::Duration::from_secs(30),
-            old_version: 1,
-            new_version: 2,
-        };
-
-        let desc = event.description();
-        assert!(desc.contains(&cred_id.to_string()));
-        assert!(desc.contains("v1"));
-        assert!(desc.contains("v2"));
-    }
-
-    struct MockSender {
-        should_fail: bool,
-    }
-
-    impl NotificationSender for MockSender {
-        async fn send(&self, event: &NotificationEvent) -> RotationResult<()> {
-            if self.should_fail {
-                Err(super::super::error::RotationError::NotificationFailed {
-                    credential_id: event.credential_id().to_string(),
-                    reason: "Mock failure".to_string(),
-                })
-            } else {
-                Ok(())
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_send_notification_success() {
-        use nebula_resilience::retry::{BackoffConfig, JitterConfig, RetryConfig};
-
-        let sender = MockSender { should_fail: false };
-        let event = NotificationEvent::RotationStarting {
-            credential_id: test_cred_id(),
-            starting_at: Utc::now(),
-            transaction_id: "tx-123".to_string(),
-        };
-
-        let config = RetryConfig::new(6)
-            .unwrap()
-            .backoff(BackoffConfig::Exponential {
-                base: std::time::Duration::from_millis(100),
-                multiplier: 2.0,
-                max: std::time::Duration::from_secs(32),
-            })
-            .jitter(JitterConfig::Full {
-                factor: 0.25,
-                seed: None,
-            });
-        let result = send_notification(&sender, &event, config).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_send_notification_failure() {
-        use nebula_resilience::retry::{BackoffConfig, RetryConfig};
-
-        let sender = MockSender { should_fail: true };
-        let event = NotificationEvent::RotationStarting {
-            credential_id: test_cred_id(),
-            starting_at: Utc::now(),
-            transaction_id: "tx-123".to_string(),
-        };
-
-        // Use short delays so test doesn't actually wait 30s.
-        let config = RetryConfig::new(3)
-            .unwrap()
-            .backoff(BackoffConfig::Exponential {
-                base: std::time::Duration::from_millis(10),
-                multiplier: 2.0,
-                max: std::time::Duration::from_millis(100),
-            });
-        let result = send_notification(&sender, &event, config).await;
-        assert!(result.is_err());
-    }
-}
