+++ crates/credential/src/service/mod.rs
@@ -6,13 +6,30 @@
 //! construction path is the only path; the composition root
 //! (`nebula-api`'s credential builder) is the only place that calls
 //! `CredentialService::from_secure_parts`.
+/// Acquisition (`resolve` / `continue_resolve`) methods of
+/// `CredentialService` (split from `facade` for size; behaviour-preserving
+/// `impl` block).
+mod acquire;
 pub mod binding;
+/// Capability operations (`test` / `refresh` / `revoke`) of
+/// `CredentialService` (split from `facade` for size; behaviour-preserving
+/// `impl` block).
+mod capabilities;
+/// CRUD (create/read/list/update/delete) methods of `CredentialService`
+/// (split from `facade` for size; behaviour-preserving `impl` block).
+mod crud;
+/// Type-discovery methods of `CredentialService` (split from `facade` for
+/// size; behaviour-preserving `impl` block).
+mod discovery;
 pub mod error;
 pub mod facade;
 pub mod head;
 pub mod observer;
 pub mod ops;
 pub mod scope;
+/// Slot / binding resolution methods of `CredentialService` (split from
+/// `facade` for size; behaviour-preserving `impl` block).
+mod slot;
 pub mod state_source;
 
 pub use binding::{TenantFingerprint, ValidatedCredentialBinding, ValidatedCredentialBindingError};
