--- crates/credential/tests/probes/capabilities_declared_without_subtrait_impl.stderr
@@ -1,22 +1,0 @@
-error[E0277]: the trait bound `DummyCredential: Refreshable` is not satisfied
-  --> tests/probes/capabilities_declared_without_subtrait_impl.rs:34:8
-   |
-34 | struct DummyCredential;
-   |        ^^^^^^^^^^^^^^^ unsatisfied trait bound
-   |
-help: the trait `Refreshable` is not implemented for `DummyCredential`
-  --> tests/probes/capabilities_declared_without_subtrait_impl.rs:34:1
-   |
-34 | struct DummyCredential;
-   | ^^^^^^^^^^^^^^^^^^^^^^
-help: the trait `Refreshable` is implemented for `OAuth2Credential`
-  --> src/credentials/oauth2.rs
-   |
-   | #[nebula_credential::credential(key = "oauth2")]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-note: required by a bound in `assert_capability`
-  --> tests/probes/capabilities_declared_without_subtrait_impl.rs:26:10
-   |
-26 | #[derive(nebula_credential_macros::Credential)]
-   |          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `assert_capability`
-   = note: this error originates in the attribute macro `nebula_credential::credential` which comes from the expansion of the derive macro `nebula_credential_macros::Credential` (in Nightly builds, run with -Z macro-backtrace for more info)
